refactor IssnGeneration to own class
